getWorkspaces and getProjects hooks seperated from the form
